move to utility class
